+++ vulkanalia/src/vk/extensions.rs
@@ -534,8 +534,13 @@ pub trait ExtBufferDeviceAddressExtension: DeviceV1_0 {
 
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkGetBufferDeviceAddressEXT.html>
     #[inline]
-    unsafe fn get_buffer_device_address_ext(&self, info: &BufferDeviceAddressInfo) {
+    unsafe fn get_buffer_device_address_ext(
+        &self,
+        info: &BufferDeviceAddressInfo,
+    ) -> DeviceAddress {
         let __result = (self.commands().get_buffer_device_address_ext)(self.handle(), info);
+
+        __result
     }
 }
 
@@ -1074,7 +1079,7 @@ pub trait ExtDirectfbSurfaceExtension: InstanceV1_0 {
         physical_device: PhysicalDevice,
         queue_family_index: u32,
         dfb: *mut IDirectFB,
-    ) {
+    ) -> Bool32 {
         let __result = (self
             .commands()
             .get_physical_device_direct_fb_presentation_support_ext)(
@@ -1082,6 +1087,8 @@ pub trait ExtDirectfbSurfaceExtension: InstanceV1_0 {
             queue_family_index,
             dfb,
         );
+
+        __result
     }
 }
 
@@ -3569,11 +3576,13 @@ pub trait KhrAccelerationStructureExtension: DeviceV1_0 {
     unsafe fn get_acceleration_structure_device_address_khr(
         &self,
         info: &AccelerationStructureDeviceAddressInfoKHR,
-    ) {
+    ) -> DeviceAddress {
         let __result =
             (self
                 .commands()
                 .get_acceleration_structure_device_address_khr)(self.handle(), info);
+
+        __result
     }
 
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkGetDeviceAccelerationStructureCompatibilityKHR.html>
@@ -3711,14 +3720,21 @@ pub trait KhrBufferDeviceAddressExtension: DeviceV1_0 {
 
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkGetBufferDeviceAddressKHR.html>
     #[inline]
-    unsafe fn get_buffer_device_address_khr(&self, info: &BufferDeviceAddressInfo) {
+    unsafe fn get_buffer_device_address_khr(
+        &self,
+        info: &BufferDeviceAddressInfo,
+    ) -> DeviceAddress {
         let __result = (self.commands().get_buffer_device_address_khr)(self.handle(), info);
+
+        __result
     }
 
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkGetBufferOpaqueCaptureAddressKHR.html>
     #[inline]
-    unsafe fn get_buffer_opaque_capture_address_khr(&self, info: &BufferDeviceAddressInfo) {
+    unsafe fn get_buffer_opaque_capture_address_khr(&self, info: &BufferDeviceAddressInfo) -> u64 {
         let __result = (self.commands().get_buffer_opaque_capture_address_khr)(self.handle(), info);
+
+        __result
     }
 
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkGetDeviceMemoryOpaqueCaptureAddressKHR.html>
@@ -3726,9 +3742,11 @@ pub trait KhrBufferDeviceAddressExtension: DeviceV1_0 {
     unsafe fn get_device_memory_opaque_capture_address_khr(
         &self,
         info: &DeviceMemoryOpaqueCaptureAddressInfo,
-    ) {
+    ) -> u64 {
         let __result =
             (self.commands().get_device_memory_opaque_capture_address_khr)(self.handle(), info);
+
+        __result
     }
 }
 
@@ -3948,9 +3966,14 @@ pub trait KhrDeferredHostOperationsExtension: DeviceV1_0 {
 
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkGetDeferredOperationMaxConcurrencyKHR.html>
     #[inline]
-    unsafe fn get_deferred_operation_max_concurrency_khr(&self, operation: DeferredOperationKHR) {
+    unsafe fn get_deferred_operation_max_concurrency_khr(
+        &self,
+        operation: DeferredOperationKHR,
+    ) -> u32 {
         let __result =
             (self.commands().get_deferred_operation_max_concurrency_khr)(self.handle(), operation);
+
+        __result
     }
 
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkGetDeferredOperationResultKHR.html>
@@ -6130,13 +6153,15 @@ pub trait KhrRayTracingPipelineExtension: DeviceV1_0 {
         pipeline: Pipeline,
         group: u32,
         group_shader: ShaderGroupShaderKHR,
-    ) {
+    ) -> DeviceSize {
         let __result = (self.commands().get_ray_tracing_shader_group_stack_size_khr)(
             self.handle(),
             pipeline,
             group,
             group_shader,
         );
+
+        __result
     }
 }
 
@@ -6985,7 +7010,7 @@ pub trait KhrWaylandSurfaceExtension: InstanceV1_0 {
         physical_device: PhysicalDevice,
         queue_family_index: u32,
         display: *mut wl_display,
-    ) {
+    ) -> Bool32 {
         let __result = (self
             .commands()
             .get_physical_device_wayland_presentation_support_khr)(
@@ -6993,6 +7018,8 @@ pub trait KhrWaylandSurfaceExtension: InstanceV1_0 {
             queue_family_index,
             display,
         );
+
+        __result
     }
 }
 
@@ -7042,13 +7069,15 @@ pub trait KhrWin32SurfaceExtension: InstanceV1_0 {
         &self,
         physical_device: PhysicalDevice,
         queue_family_index: u32,
-    ) {
+    ) -> Bool32 {
         let __result = (self
             .commands()
             .get_physical_device_win32_presentation_support_khr)(
             physical_device,
             queue_family_index,
         );
+
+        __result
     }
 }
 
@@ -7100,7 +7129,7 @@ pub trait KhrXcbSurfaceExtension: InstanceV1_0 {
         queue_family_index: u32,
         connection: *mut xcb_connection_t,
         visual_id: xcb_visualid_t,
-    ) {
+    ) -> Bool32 {
         let __result = (self
             .commands()
             .get_physical_device_xcb_presentation_support_khr)(
@@ -7109,6 +7138,8 @@ pub trait KhrXcbSurfaceExtension: InstanceV1_0 {
             connection,
             visual_id,
         );
+
+        __result
     }
 }
 
@@ -7150,7 +7181,7 @@ pub trait KhrXlibSurfaceExtension: InstanceV1_0 {
         physical_device: PhysicalDevice,
         queue_family_index: u32,
         visual_id: VisualID,
-    ) -> Display {
+    ) -> (Bool32, Display) {
         let mut dpy = MaybeUninit::<Display>::uninit();
 
         let __result = (self
@@ -7162,7 +7193,7 @@ pub trait KhrXlibSurfaceExtension: InstanceV1_0 {
             visual_id,
         );
 
-        dpy.assume_init()
+        (__result, dpy.assume_init())
     }
 }
 
@@ -7399,8 +7430,10 @@ pub trait NvxImageViewHandleExtension: DeviceV1_0 {
 
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkGetImageViewHandleNVX.html>
     #[inline]
-    unsafe fn get_image_view_handle_nvx(&self, info: &ImageViewHandleInfoNVX) {
+    unsafe fn get_image_view_handle_nvx(&self, info: &ImageViewHandleInfoNVX) -> u32 {
         let __result = (self.commands().get_image_view_handle_nvx)(self.handle(), info);
+
+        __result
     }
 }
 
@@ -8569,7 +8602,7 @@ pub trait QnxScreenSurfaceExtension: InstanceV1_0 {
         physical_device: PhysicalDevice,
         queue_family_index: u32,
         window: *mut _screen_window,
-    ) {
+    ) -> Bool32 {
         let __result = (self
             .commands()
             .get_physical_device_screen_presentation_support_qnx)(
@@ -8577,6 +8610,8 @@ pub trait QnxScreenSurfaceExtension: InstanceV1_0 {
             queue_family_index,
             window,
         );
+
+        __result
     }
 }